requirements.txt
python-dotenv
google-adk
google-adk[a2a]
uvicorn